+++ src/config.rs
@@ -1,7 +1,7 @@
 use crate::compressor;
 use crate::filters;
 use crate::mixer;
-use crate::wavtools::find_data_in_wav_stream;
+use crate::wavtools::{find_data_in_wav_stream, WavParams};
 use parking_lot::RwLock;
 use serde::{de, Deserialize, Serialize};
 //use serde_with;
@@ -17,6 +17,7 @@ use std::path::{Path, PathBuf};
 use crate::PrcFmt;
 type Res<T> = Result<T, Box<dyn error::Error>>;
 
+#[derive(Clone)]
 pub struct Overrides {
     pub samplerate: Option<usize>,
     pub sample_format: Option<SampleFormat>,
@@ -203,8 +204,8 @@ pub enum CaptureDevice {
         device: String,
         format: SampleFormat,
     },
-    #[serde(alias = "FILE", alias = "file")]
-    File(CaptureDeviceFile),
+    RawFile(CaptureDeviceRawFile),
+    WavFile(CaptureDeviceWavFile),
     #[serde(alias = "STDIN", alias = "stdin")]
     Stdin(CaptureDeviceStdin),
     #[cfg(target_os = "macos")]
@@ -245,7 +246,10 @@ impl CaptureDevice {
             CaptureDevice::Bluez(dev) => dev.channels,
             #[cfg(feature = "pulse-backend")]
             CaptureDevice::Pulse { channels, .. } => *channels,
-            CaptureDevice::File(dev) => dev.channels,
+            CaptureDevice::RawFile(dev) => dev.channels,
+            CaptureDevice::WavFile(dev) => {
+                dev.wav_info().map(|info| info.channels).unwrap_or_default()
+            }
             CaptureDevice::Stdin(dev) => dev.channels,
             #[cfg(target_os = "macos")]
             CaptureDevice::CoreAudio(dev) => dev.channels,
@@ -269,12 +273,11 @@ impl CaptureDevice {
 
 #[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
 #[serde(deny_unknown_fields)]
-pub struct CaptureDeviceFile {
+pub struct CaptureDeviceRawFile {
     #[serde(deserialize_with = "validate_nonzero_usize")]
     pub channels: usize,
     pub filename: String,
-    #[serde(default)]
-    pub format: Option<SampleFormat>,
+    pub format: SampleFormat,
     #[serde(default)]
     pub extra_samples: Option<usize>,
     #[serde(default)]
@@ -283,7 +286,7 @@ pub struct CaptureDeviceFile {
     pub read_bytes: Option<usize>,
 }
 
-impl CaptureDeviceFile {
+impl CaptureDeviceRawFile {
     pub fn extra_samples(&self) -> usize {
         self.extra_samples.unwrap_or_default()
     }
@@ -292,6 +295,33 @@ impl CaptureDeviceFile {
     }
     pub fn read_bytes(&self) -> usize {
         self.read_bytes.unwrap_or_default()
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
+#[serde(deny_unknown_fields)]
+pub struct CaptureDeviceWavFile {
+    pub filename: String,
+    #[serde(default)]
+    pub extra_samples: Option<usize>,
+}
+
+impl CaptureDeviceWavFile {
+    pub fn extra_samples(&self) -> usize {
+        self.extra_samples.unwrap_or_default()
+    }
+
+    pub fn wav_info(&self) -> Res<WavParams> {
+        let fname = &self.filename;
+        let f = match File::open(fname) {
+            Ok(f) => f,
+            Err(err) => {
+                let msg = format!("Could not open input file '{fname}'. Reason: {err}");
+                return Err(ConfigError::new(&msg).into());
+            }
+        };
+        let file = BufReader::new(&f);
+        find_data_in_wav_stream(file)
     }
 }
 
@@ -1317,7 +1347,7 @@ pub fn load_config(filename: &str) -> Res<Configuration> {
     let file = match File::open(filename) {
         Ok(f) => f,
         Err(err) => {
-            let msg = format!("Could not open config file '{filename}'. Error: {err}");
+            let msg = format!("Could not open config file '{filename}'. Reason: {err}");
             return Err(ConfigError::new(&msg).into());
         }
     };
@@ -1326,7 +1356,7 @@ pub fn load_config(filename: &str) -> Res<Configuration> {
     let _number_of_bytes: usize = match buffered_reader.read_to_string(&mut contents) {
         Ok(number_of_bytes) => number_of_bytes,
         Err(err) => {
-            let msg = format!("Could not read config file '{filename}'. Error: {err}");
+            let msg = format!("Could not read config file '{filename}'. Reason: {err}");
             return Err(ConfigError::new(&msg).into());
         }
     };
@@ -1337,16 +1367,24 @@ pub fn load_config(filename: &str) -> Res<Configuration> {
             return Err(ConfigError::new(&msg).into());
         }
     };
-    //Ok(configuration)
-    //apply_overrides(&mut configuration);
-    //replace_tokens_in_config(&mut configuration);
-    //replace_relative_paths_in_config(&mut configuration, filename);
     Ok(configuration)
 }
 
 fn apply_overrides(configuration: &mut Configuration) {
-    // TODO update to get overrides from wav capture file
-    let overrides = OVERRIDES.read();
+    let mut overrides = OVERRIDES.read().clone();
+    // Only one match arm for now, might be more later.
+    #[allow(clippy::single_match)]
+    match &configuration.devices.capture {
+        CaptureDevice::WavFile(dev) => {
+            if let Ok(wav_info) = dev.wav_info() {
+                overrides.channels = Some(wav_info.channels);
+                overrides.sample_format = Some(wav_info.sample_format);
+                overrides.samplerate = Some(wav_info.sample_rate);
+                debug!("Updating overrides with values from wav input file, rate {}, format: {}, channels: {}", wav_info.sample_rate, wav_info.sample_format, wav_info.channels);
+            }
+        }
+        _ => {}
+    }
     if let Some(rate) = overrides.samplerate {
         let cfg_rate = configuration.devices.samplerate;
         let cfg_chunksize = configuration.devices.chunksize;
@@ -1366,7 +1404,7 @@ fn apply_overrides(configuration: &mut Configuration) {
             configuration.devices.chunksize = scaled_chunksize;
             #[allow(unreachable_patterns)]
             match &mut configuration.devices.capture {
-                CaptureDevice::File(dev) => {
+                CaptureDevice::RawFile(dev) => {
                     let new_extra = dev.extra_samples() * rate / cfg_rate;
                     debug!(
                         "Scale extra samples: {} -> {}",
@@ -1399,7 +1437,7 @@ fn apply_overrides(configuration: &mut Configuration) {
         debug!("Apply override for extra_samples: {}", extra);
         #[allow(unreachable_patterns)]
         match &mut configuration.devices.capture {
-            CaptureDevice::File(dev) => {
+            CaptureDevice::RawFile(dev) => {
                 dev.extra_samples = Some(extra);
             }
             CaptureDevice::Stdin(dev) => {
@@ -1411,9 +1449,10 @@ fn apply_overrides(configuration: &mut Configuration) {
     if let Some(chans) = overrides.channels {
         debug!("Apply override for capture channels: {}", chans);
         match &mut configuration.devices.capture {
-            CaptureDevice::File(dev) => {
+            CaptureDevice::RawFile(dev) => {
                 dev.channels = chans;
             }
+            CaptureDevice::WavFile(_dev) => {}
             CaptureDevice::Stdin(dev) => {
                 dev.channels = chans;
             }
@@ -1458,10 +1497,10 @@ fn apply_overrides(configuration: &mut Configuration) {
     if let Some(fmt) = overrides.sample_format {
         debug!("Apply override for capture sample format: {}", fmt);
         match &mut configuration.devices.capture {
-            CaptureDevice::File(dev) => {
-                // TODO dont update if null!
-                dev.format = Some(fmt);
+            CaptureDevice::RawFile(dev) => {
+                dev.format = fmt;
             }
+            CaptureDevice::WavFile(_dev) => {}
             CaptureDevice::Stdin(dev) => {
                 dev.format = fmt;
             }
@@ -1807,28 +1846,30 @@ pub fn validate_config(conf: &mut Configuration, filename: Option<&str>) -> Res<
             .into());
         }
     }
-    if let CaptureDevice::File(dev) = &conf.devices.capture {
+    if let CaptureDevice::RawFile(dev) = &conf.devices.capture {
+        let fname = &dev.filename;
+        match File::open(fname) {
+            Ok(f) => f,
+            Err(err) => {
+                let msg = format!("Could not open input file '{fname}'. Reason: {err}");
+                return Err(ConfigError::new(&msg).into());
+            }
+        };
+    }
+    if let CaptureDevice::WavFile(dev) = &conf.devices.capture {
         let fname = &dev.filename;
         let f = match File::open(fname) {
             Ok(f) => f,
             Err(err) => {
-                let msg = format!("Could not open input file '{fname}'. Error: {err}");
+                let msg = format!("Could not open input file '{fname}'. Reason: {err}");
                 return Err(ConfigError::new(&msg).into());
             }
         };
-        if dev.format.is_none() {
-            let file = BufReader::new(&f);
-            let wav_info = find_data_in_wav_stream(file)?;
-            // TODO warn if samplerate is wrong
-            // TODO 2, next step: update overrides for channels and sample rate
-            if wav_info.channels != dev.channels {
-                let msg = format!(
-                    "Wav file '{}' has wrong number of channels, expected: {}, found: {}",
-                    fname, dev.channels, wav_info.channels
-                );
-                return Err(ConfigError::new(&msg).into());
-            }
-        }
+        let file = BufReader::new(&f);
+        let _wav_info = find_data_in_wav_stream(file).map_err(|err| {
+            let msg = format!("Error reading wav file '{fname}'. Reason: {err}");
+            ConfigError::new(&msg)
+        })?;
     }
     let mut num_channels = conf.devices.capture.channels();
     let fs = conf.devices.samplerate;